Allow large Request-URIs
This is a workaround for a questionable design decision in the EidoGo widget
(namely, to pass the entire SGF to be downloaded as a query parameter). Until
a larger-scale refactoring takes place, this at least dramatically raises the
barrier before the Request-URI limit gets hit.

As a scale reference, the large SGF used as an example in the bug report was
12.8K and this patch raises the limit to 32K, nearly 3x that size.